set max_seq_length to Int64

+++ src/transformer/cent_select/ophost/cent_select_def.cpp
@@ -57,7 +57,7 @@ class CentSelect : public OpDef {
                 .UnknownShapeFormat({ge::FORMAT_ND});  
             this->Output("max_page_position_length")
                 .ParamType(REQUIRED)
-                .DataType({ge::DT_INT32})
+                .DataType({ge::DT_INT64})
                 .Format({ge::FORMAT_ND})
                 .UnknownShapeFormat({ge::FORMAT_ND});
             // this->Output("indices")
